UPIbackend/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
executing==2.0.1
fonttools==4.50.0
gunicorn==23.0.0
idna==3.6
ipython==8.22.2
jedi==0.19.1
kiwisolver==1.4.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
numpy==1.26.4
packaging==24.0
parso==0.8.3
pillow==10.2.0
prompt-toolkit==3.0.43
pure-eval==0.2.2
Pygments==2.17.2
PyJWT==2.9.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.31.0
six==1.16.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.2
tzdata==2025.2
urllib3==2.2.1
wcwidth==0.2.13
